% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concConvert.R
\name{concConvert}
\alias{concConvert}
\title{Convert a table of peak intensities to concentrations}
\usage{
concConvert(exCofit, inttable)
}
\arguments{
\item{exCofit}{the output of an nls function finding the extinction coefficients}

\item{inttable}{the table of intensities used in the previous nls function}
}
\description{
Uses the output of exConls and the original table of intensities to give concentration. Take care with units!
}

