% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concPlot.R
\name{concPlot}
\alias{concPlot}
\title{Plot fitted concentrations against the total concentration}
\usage{
concPlot(conctable, concentration, times, plotTitle)
}
\arguments{
\item{conctable}{a data frame of concentrations}

\item{concentration}{the specified total concentration}

\item{times}{the timepoints of your concentrations}

\item{plotTitle}{The title you want on your plot}
}
\description{
Plots the concentrations calculated by concConvert and exConls, adding lines for the defined concentration against the total concentration of the curve.
}

