% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conLtoM.R
\name{conLtoM}
\alias{conLtoM}
\title{Get a matrix of constants from mkin}
\usage{
conLtoM(conList, condition)
}
\arguments{
\item{conList}{a list of mkin outputs}

\item{condition}{a vector containing the conditions you want to split the coefficients by in the order they are in your list}
}
\description{
Turn a list of mkin outputs into a matrix of the coefficients
}

