% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constantsout.R
\name{constantsOut}
\alias{constantsOut}
\title{Gets the constants from mkin output}
\usage{
constantsOut(model)
}
\arguments{
\item{model}{mkin output from which the coefficients export}
}
\description{
This is pointless except as a part of another function
}
\keyword{internal}

