% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estOut.R
\name{estOut}
\alias{estOut}
\title{Collate a list of mkin bpars into a table}
\usage{
estOut(outList, conditions, row)
}
\arguments{
\item{outList}{a list of mkin bpars stored in data frames, output of summaryOut}

\item{conditions}{a vector of the experimental conditions the items of your outList are separated by}

\item{row}{an integer specifying the row of the tables within outList you wish to export}
}
\description{
The function summaryOut extracts the coefficients and confidence intervals from mkin output. This takes a list of these outputs and collates them in a matrix
}

