% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exConls.R
\name{exConls}
\alias{exConls}
\title{Fit extinction coefficients to your intensity data}
\usage{
exConls(conc, inttable, magnitude)
}
\arguments{
\item{conc}{The total concentration of your species.}

\item{inttable}{A dataframe with columns containing the intensities of each peak across your dataset.}

\item{magnitude}{An rough guess for the extinction coefficients. This assumes all are of a similar order of magnitude.}
}
\value{
Extinction coefficients for each column. Named exCo_ for the number of the column. see \code{\link[stats]{nls}}
}
\description{
Uses nls to fit pseudo-extinction coefficients to three intensities, given total concentration. If you want to do any number but three, you'll need a new function.
}

