% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exCoYeah.R
\name{exCoYeah}
\alias{exCoYeah}
\title{Calculate the concentrations of species from intensities}
\usage{
exCoYeah(concentration, inttable, times, plotTitle)
}
\arguments{
\item{concentration}{the known total concentration}

\item{inttable}{a table in which each column is a set of intensities for a peak}

\item{times}{the timepoints at which you collected spectra}

\item{plotTitle}{the title you want on the plot this delivers}
}
\description{
From a given concentration, calculates extinction coefficients and concentrations for three sets of peak intensities. Also plots the concentrations against time. The name of the function is stupid, I just couldn't think of a better one
}

