% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAllPeaks.R
\name{findAllPeaks}
\alias{findAllPeaks}
\title{Finds the local maxima for a list of fitted experiments}
\usage{
findAllPeaks(bigList, posTable)
}
\arguments{
\item{bigList}{A list containing lists of data frames of spectra with columns "absorbance" and "wavelength"}

\item{posTable}{A data frame or matrix containing the centres and widths of peaks. Columns should alternate as the centre and width of a peak, with the centre first.}
}
\description{
Given a list of lists of spectra (i.e. each item of the list is a list of spectra as generated by \code{\link[psRIIfit]{spect.frametolist}}) and a table of peak parameters, finds the true maxima within the peaks' ranges.
}

