% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findLocalMax.R
\name{findLocalMax}
\alias{findLocalMax}
\title{Finds the maximum of a spectrum within a peak's width}
\usage{
findLocalMax(centre, width, spectrum, wavmax)
}
\arguments{
\item{centre}{The centre of the peak.}

\item{width}{The FWHM of the peak.}

\item{spectrum}{The spectrum for inspection.}

\item{wavmax}{The maximum wavelength of the spectrum (largest wavelength).}
}
\description{
When given a peak's centre and full width at half maximum, can find the position of the maximum within that range for a given spectrum
}
\keyword{internal}

