% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDemPeaks.R
\name{fitDemPeaks}
\alias{fitDemPeaks}
\title{A formula to fit peaks to a set of data}
\usage{
fitDemPeaks(fixed, guesses, spectra, wavelengths, belowZero = FALSE)
}
\arguments{
\item{fixed}{A data frame containing columns named "centres" and "widths", specifying the centre and HWHM of each peak respectively.}

\item{guesses}{A list of starting guesses for your data. nls.LM is smart enough that these usually just need to be of the right order of magnitude. In a list, called "i1", "i2" etc.}

\item{spectra}{A list of data frames containing wavelength and absorbance data.}

\item{wavelengths}{A vector of the wavelengths your data is over.}

\item{belowZero}{An optional logical value. Defaults to false, meaning intensities bottom out at zero.}
}
\value{
A matrix of intensities for each peak in fixed.
}
\description{
Fits a set of peaks with defined centres and widths (FWHM) to a range of spectra. I know the name is silly.
}

