% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussgen.R
\name{gaussgen}
\alias{gaussgen}
\title{A generator of multiple gaussian peaks}
\usage{
gaussgen(peakNumber, fixedParameters, wavelength)
}
\arguments{
\item{peakNumber}{The number of peaks you want to generate}

\item{fixedParameters}{a list of fixed centres and half widths at half maximum. Centres are named "c1 = " etc., and widths as "w1 = " etc.}

\item{wavelength}{the wavelength of the dataset you want to fit the peaks to}
}
\description{
Generates up to four gaussian peaks (I was too lazy to make it more general)
}

