% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussint.R
\name{gaussint}
\alias{gaussint}
\title{Intensity gaussian}
\usage{
gaussint(intensity, wavelength, centre, width)
}
\arguments{
\item{intensity}{maximum height of the peak}

\item{wavelength}{the x-axis of the peak (this is a very niche package)}

\item{centre}{the centre of the peak}

\item{width}{the full width at half maximum}
}
\value{
A gaussian peak
}
\description{
This function generates a gaussian function
}

