% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxFinder.R
\name{maxFinder}
\alias{maxFinder}
\title{Find the local maxima for the most representative spectrum}
\usage{
maxFinder(peakList, fitList)
}
\arguments{
\item{peakList}{A list of local maxima from \code{\link[psRIIfit]{findAllPeaks}}}

\item{fitList}{A list of intensity fits from \code{\link[psRIIfit]{fitDemPeaks}}.}
}
\description{
Takes the local maxima of peak ranges for a dataset (e.g. timepoints) or a list thereof and the intensities of those peaks fitted to the dataset and finds the most representative maximum.
}

