% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkinFormat.R
\name{mkinFormat}
\alias{mkinFormat}
\title{Converts concentrations into mkin's format}
\usage{
mkinFormat(conctable, times)
}
\arguments{
\item{conctable}{a table containing three columns of concentrations}

\item{times}{a vector of the timepoints at which spectra were collected}
}
\description{
takes a data table of concentrations and a vector of times into mkin's slightly weird format
}

