% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realPeakFinder.R
\name{realPeakFinder}
\alias{realPeakFinder}
\title{Find the true maxima of spectra within peak ranges}
\usage{
realPeakFinder(spectraList, centres, widths)
}
\arguments{
\item{spectraList}{a list of data frames containing absorbance and wavelength columns. The wavelengths must be integer values from largest to smallest.}

\item{centres}{a vector of the centres of the peaks modeled, given as wavelengths.}

\item{widths}{a vector of the widths of the peaks modeled, given as wavelengths.}
}
\description{
When given a list of spectra and vectors of peak parameters, returns the maxima within those peak ranges for each spectrum
}

