% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameFrame.R
\name{renameFrame}
\alias{renameFrame}
\title{Rename data frames to be compatible with fitDemPeaks}
\usage{
renameFrame(df)
}
\arguments{
\item{df}{a dataframe in which the first column contains wavelength and the second absorbance}
}
\description{
This renames two columns of a data frame to "wavelength" and "absorbance"
}
\keyword{internal}

