% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortspec.R
\name{shortspec}
\alias{shortspec}
\title{Shorten spectra to your desired length}
\usage{
shortspec(spectrum, length)
}
\arguments{
\item{spectrum}{a data frame or matrix containing your data}

\item{length}{how long you want it to end up}
}
\description{
I was too lazy to do this a bunch of times. Just cuts the end off, nothing fancy
}

