% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryOut.R
\name{summaryOut}
\alias{summaryOut}
\title{Export coefficients from mkin summary}
\usage{
summaryOut(mkinOutput)
}
\arguments{
\item{mkinOutput}{The original mkin output}
}
\description{
The "bpar" of an mkin list contains the coefficients and their confidence intervals. This takes these and puts them in a table
}

