% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exConls.R
\name{exConls}
\alias{exConls}
\title{Fit extinction coefficients to your intensity data}
\usage{
exConls(concentration, inttable)
}
\arguments{
\item{concentration}{the total concentration of your species}

\item{inttable}{a dataframe containing three columns, each containing the intensity of your peaks across a dataset}
}
\description{
Uses nls to fit pseudo-extinction coefficients to three intensities, given total concentration. If you want to do any number but three, you'll need a new function.
}

