% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDemPeaks.R
\name{fitDemPeaks}
\alias{fitDemPeaks}
\title{A formula to fit peaks to a set of data}
\usage{
fitDemPeaks(peaks, fixed, guesses, spectra, wavelengths)
}
\arguments{
\item{peaks}{The number of peaks you wish to fit (currently no more than four)}

\item{fixed}{a list of defined centres and widths for your peaks. See the documentation for gaussgen for format}

\item{guesses}{a list of starting guesses for your data. nls.LM is smart enough that these usually just need to be of the right order of magnitude. In a list, called "i1", "i2" etc.}

\item{wavelengths}{a vector of the wavelengths your data is over}
}
\description{
Fits a set of peaks with defined centres and widths (HWHM) to a range of spectra. I know the name is silly.
}

