% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussgen.R
\name{gaussgen}
\alias{gaussgen}
\title{A generator of multiple gaussian peaks}
\usage{
gaussgen(fixedParameters, wavelengths)
}
\arguments{
\item{fixedParameters}{A data frame containing columns named "centres" and "widths". Each row contains the parameters of one peak.}

\item{wavelength}{The wavelengths of the dataset you want to fit the peaks to.}
}
\value{
A formula for the number of peaks specified by the number of rows in fixedParameters
}
\description{
Generates any number of gaussian peaks of indeterminate
}
\keyword{internal}

