% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameFrame.R
\name{renameFrame}
\alias{renameFrame}
\title{Rename data frames to be compatible with fitDemPeaks}
\usage{
renameFrame(df)
}
\description{
This renames two columns of a data frame to "wavelength" and "absorbance"
}

