% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectframetolist.R
\name{spect.frametolist}
\alias{spect.frametolist}
\title{Convert a data frame of spectra to a list}
\usage{
spect.frametolist(df)
}
\description{
Converts a data frame with the wavelengths and recorded spectra into a list containing data frames of wavelength and recorded spectra, compatible with fitDemPeaks.
}

